Bump WasmBuildTests timeout to 180min
Looking at the data for the last 30 days a successful or failed WasmBuildTests job took between 70-90 minutes.
When Helix queues are overloaded the 120min timeout can result in canceled builds which is especially true on servicing queues which first need to spin up VMs so bump the timeout to 180mins.