Fix generating debug info for static fields (#95570)
Fixes #93868.

The existing algorithm for generating debug info about static fields was:

* If we're generating debug information about a type, generate debug information about all non-literal and non-RVA fields on it.

This is both too little, and too much.

It's too little: if the only thing we're touching on a type is reading/writing the static fields, we'd not be generating debug information for that type because the current logic only looks at methods and `MethodTable`s.

It's too much: if we didn't touch any statics on the type, the static base wasn't generated either. Debug information would refer to something that doesn't exist.

I'm updating this logic to look at actual bases instead - if a static/gcstatic/threadstatic/instance base was generated, generate debug info for the fields. Skip them otherwise because they got optimized away ("trimmed").