JIT: Add post-phase validation of DFS tree
Add a post-phase check that validates that the DFS tree matches the
current flow graph when it is non-null.

Add a `fgInvalidateDfsTree()` that invalidates the DFS tree and derived
annotations on it (dominators and loops). We invalidate after RBO, which
is the last point we expect to make use of the flow graph information.

Compute new DFS and loops as part of loop unrolling to make the DFS tree
correct after this point too. Some TP regressions are expected due to
this.

To sum up we now keep this information valid from loop finding until
RBO.

Fix #95652