[mono][interp] Implement context register backup and restore mechanism
This commit adds backup and restore ins for all context registers. In the first part, the tramp backups all context registers by storing them in the ccontext. In the second part, the tramp restores all context registers by loading them from the ccontext. This is necessary for maintaining the execution context after the native call.