JIT: more CSE refactoring
Refactor so that the heuristic is also responsible for the initial
candidate selection. This will allow us to experiment with what
sorts of trees can be CSEs.

Pre-CSE queries now are also routed  through the heuristic. This
leads to a few diffs in cases where hosting and CSE did not always
agree on what could be CSEd.