Added guard to throw InvalidOperrationException.
Added guard to throw InvalidOperrationException if the argument typeToConvert and Converter are incompatible.
The exception message is "Type 'NameOfType' is not compatible with converter 'ConverterName'." and "Type 'NameOfType' is not compatible with converter 'ConverterName'.

NOTES:
The test threw a NotSupportedException on the following line and could not reproduce this issue, so the JsonSerializerOptions instantiation was changed.
OptionsTests.ConverterRead_VerifyInvalidTypeToConvertFails() lineno 1432
KeyValuePair<int, int> kvp = converter.Read(ref reader, typeToConvert, options);

Fix #36605