JIT: Make isRemovableJmpCandidate less restrictive on AMD64 (#95493)
Previously, the emitter-level jump removal optimization would not consider jumps following call instructions on AMD64, as unwind semantics require there to be an instruction between a call and an OS epilogue. This was overly restrictive, as many call-jmp pairs do not precede an OS epilogue, and for those that do, the jump can be replaced with a nop. This change allows jumps following calls to be considered for removal, and inserts a nop if needed.