[mono][interp] Refactor context register backup and restore mechanism
This commit updates backup and restore mechanism for all context registers. The native trampoline should always restore the original value of the ctx_regs to be compatible with C cconv. Before the native call, all context registers are saved onto the stack. After the native call, the context registers are saved in ccontext and then restored from the stack.