JIT: Remove most loop-related hoisting quirks (#95734)
We still keep the quirk that we only hoist in loops also recognized by
the old loop finding. Before we can hoist in new loops we need to
implement creation of preheaders, but that requires incrementally
keeping the old loop table up to date, which I'd like to avoid. So my
next goal is to switch everything to only have the "recognized by old
loop finding" quirk so that we can remove the old loop validity
requirements.

A few minor diffs are expected because the new loop iteration order is
not exactly the same as the old one. In particular child loop sibling
links are stored in rever postorder (the old loop table has them in
postorder), which means we end up hoisting them in postorder (the old
one did them in reverse postorder). I initially added an
`ArrayStack::Reverse` method to make it the same as previously, but the
diffs were so small and just constituted some reordering, so the
complication did not seem worth it.