[wasm] CI: Surface failures for MT job when runtime has changes
`runtime-wasm` has a job for running tests with multithreaded runtime,
but since that is not stable yet, it uses `shouldContinueOnError: true`.
IOW, the test failures don't cause the job to fail, and instead just
show up as `orange`. And this does not show up in github at all.

Instead, to help with devs working on the mt runtime, turn this to
`false` if there are any wasm runtime related changes.