JIT: Port loop alignment to new loop representation
* Revive Bruce's old #62940: do alignment candidate identification
  during the main alignment phase. Do this using the new loop structures
  instead.
* To avoid maintaining loop information into the backend we reidentify
  loops during the phase, so there is some TP cost. It's minimized by
  the fact that we can conservatively know whether there might be any
  loops. Also, the DFS we need to do is likely to be used in the future
  for block layout.
* Introduce `BBF_OLD_LOOP_HEADER_QUIRK`; it is a simpler version of the
  `m_newToOldLoop` array that we currently have. The next work item will
  be to switch all the loop-related quirks onto this flag instead of
  using `m_newToOldLoop`.

A few alignment related diffs expected -- sometimes from different
insertion of alignment (because the old `BasicBlock::bbNatLoopNum` was
not up-to-date), other times from more alignment because of the weight
change (see #62940 for the details on that). Also, we allow compaction
of some loop headers now that previously would be rejected from being
marked with the "align" flag, so that also causes some diffs.