null-terminate error message in swift
We must add null-terminator on the swift side to the end of the string so that it can be read correctly by PtrToStringUni on C# side (requires null-terminator).
There is a swift function that could be used https://developer.apple.com/documentation/foundation/nsstring/1415702-getcstring for this but it works with CChar not unichar.
CChar is C-char like type which is either signed or unsigned depending on the platform.
I don't think C# has a similar type to map onto so I think it is better to stick with unichar and add null-terminator manually.