Allow preinitializing `AppContext.TryGetSwitch`
Interpret this in the static constructor interpreter. Allows compile-time initializing ~30 extra types in the Stage1 app. 99% of them are the SR class.

To actually preinitialize the SR class I had to pass `--feature:System.Resources.UseSystemResourceKeys=false` on the ILC command line. This will get overwritten by `--feature:System.Resources.UseSystemResourceKeys=true` if it shows up on the command line after this.

I considered doing this in the substitution IL rewriter (so that RyuJIT-produced code can also benefit), but the API shape (returning bool and taking ref to a bool) makes this really awkward to rewrite in IL. Much easier to interpret.