checkpoint: sketch of preloading
punted on preload_visit_typespec()

will need to read a type spec blob.  Unfortunately our parser is
inexplicably tied to MonoClass construction (for MONO_TYPE_CLASS we
get the token and then call `mono_class_get_checked` which kicks off
class initialization right away) The parsing code is huge and
complicated.  I think rather than trying to copy it for the
preloader (or trying to refactor it as part of this work), a better
approach is to pass a max readiness level into the parser and into
class construction.