Make sorting of modules more stable in the compiler (#96006)
The compiler is multithreaded and to achieve stable outputs, it needs to sort everything before writing it to disk. The way we sort modules currently is by looking at the module MVID. It's a very simple, reliable, and efficient way.

Except this causes disproportionally large differences in the compiler outputs when a small part of the program was changed - changing something small in an assembly could result in large blocks of methods being shifted in the output executable. This makes delta patching very inefficient.

I'm changing the sorting to use assembly names instead. It's less efficient, but doesn't cause problems for delta patching.