extract side-effects from op1 of GT_COMMA
In `GT_COMMA(op1, op2)` `op1` is a side-effect tree, where as `op2` is considered as value and side-effect tree.
As such, we should always extract the side effects for `op1` otherwise in cases where there is a cctor call needed,
we might accidently remove the `op1` if it is not used anywhere else.