JIT: Remove BBF_NONE_QUIRK check in loop alignment (#95999)
Removes `BBF_NONE_QUIRK` check when deciding to place loop alignment padding after an unconditional jump. Instead, we now consider placing padding after a jump only if the jump won't be removed by the jump-to-next optimization. Also replaces an assert on whether `BBF_NONE_QUIRK` is set with a check that the block jumps to the next block.