JIT: Generalize check for cycles when finding loops (#96153)
`optFindNewLoops` tries to leave a breadcrumb for loop alignment about
whether it needs to bother with identifying loops. It does so by
checking for whether any natural loops were found, but also whether any
irreducible loops were found (i.e. any cycle). The idea is that
optimizations that run between finding loops and loop alignment can
easily remove edges that make irreducible loops into natural loops.

However, the check for "any cycle" was not precise enough; it made the
assumption that `NumLoops() > 0 || NumImproperLoopHeaders > 0` was
equivalent to there being a cycle in the flow graph. But that is not true,
specifically because we only consider blocks to be loop headers when
their backedges are non-exceptional. Odd loop structures can exist that
iterate through required exception throws, like in the case of #96145. In
that case we have a cycle through the handler, which ends up being a
natural loop after some optimizations run.

To fix the situation we generalize `fgComputeDfs` to identify when the
flow graph has a cycle, which is the fully general property we want for
the breadcrumb. As a bonus we can skip identifying loops if we know the
flow graph has no cycles.

Fix #96145