Fix source_file bug in mono_ppdb_lookup_location_internal
document-records affect sequence points that follow in the byte stream,
not points previously read.

When reading the document-record we may already be at the target offset.
Do not update docname local if we have reached the end target already
otherwise the wrong source_file will be reported for the requested offset