JIT: Avoid cloning cold descendant loops repeatedly
When we have a loop nest where we clone multiple ancestor loops we would
clone descendant loops, ending up with a total quadratic number of
cloned loops in terms of the depth of the loop nest.

Instead of producing these multiple versions of cold loops (each of
which is optimized for a particular hot ancestor loop) we can reuse the
first cold loop we cloned. This has the effect of exiting all the "hot"
versions of ancestor loops if the cloning conditions fail, but it gets
rid of the quadratic loop cloning problem.