JIT: Move `BBF_OLD_LOOP_HEADER_QUIRK` into old loop finding
This is currently set as part of new loop finding when we build the
old<->new loop mappings. However, the code to build these mappings
depends on `BasicBlock::bbNatLoopNum` and as we stop maintaining the old
loop table this becomes worse and worse at keeping the quirk accurate.
Instead just set it as part of old loop finding and keep it set from
then on.