Do not limit InlineArrays on Native AOT
Following program works fine with JIT, but throws a `TypeLoadException` with AOT:

```csharp
using System.Runtime.CompilerServices;

unsafe
{
    fixed (Inline* pInline = &Inline.s_inst)
        *(byte*)pInline = 123;
}

[InlineArray(640 * 480 * 4)]
struct Inline
{
    internal static Inline s_inst;
    private byte _bytes;
}
```

CoreCLR/JIT seems to have a similar block in place but only if the layout of the InlineArray is Auto, for whatever reason. Native AOT is typically more permissive. It's not clear why there is such limit on CoreCLR/JIT in the first place; we don't limit `fixed` arrays in C# this way.