[wasm] Wasm.Build.Tests - build for host os/arch instead of wasm
The tests themselves are built to run on coreclr, but each test would
build+run tests on wasm.

The CI builds are structured such that WBT gets built with
`TargetOS=browser` and `TargetArchitecture=wasm`, which would cause the
core assemblies to be bundled from the wasm runtime pack, resulting in
failures like:

```
      System.AggregateException : One or more errors occurred. (System.Diagnostics.Process is not supported on this platform.)
      ---- System.PlatformNotSupportedException : System.Diagnostics.Process is not supported on this platform.
      Stack Trace:
           at System.Threading.Tasks.Task.ThrowIfExceptional(Boolean includeTaskCanceledExceptions)
           at System.Threading.Tasks.Task`1.GetResultCore(Boolean waitCompletionNotification)
           at System.Threading.Tasks.Task`1.get_Result()
```

To make sure that WBT is built for the host os/arch we clear the two
properties.