Make it possible for arrays to have default-implemented methods
Within the managed type system, the array types are special. When asking question about virtual/interface methods implemented by arrays, one needs to switch from the array type (that doesn't have any virtual methods as far as the type system in concerned) to the `Array<T>` type from corelib (that has the methods).

The `GetClosestDefType` API is the API that does this transformation. We were not consistent enough in using it and it didn't work for default methods.

In #95830, arrays are going to get a default interface method (the implementation of `IList<T>.System.Collections.Generic.IReadOnlyList<T>.get_Item` which is `virtual final` but still a new slot).