JIT: Generalize `optSetPreheaderWeight`
Old loop finding only recognized loops with one entry block, so the old
logic here was based on there only being one entering edge to the newly
added preheader. Also, it was assumed that block was a `BBJ_COND`.

Generalize the logic by summing up the contributions to the weight from
all entering blocks. Avoid unnecessary checks on `BBJ_COND` by switching
to use the general successor visitors instead.

We only consider the preheader to have a profile weight if the following
conditions are true:
* The loop header has a profile weight
* All entering blocks have a profile weight
* The successors of all entering blocks have a profile weight (so that
  we can trust the computed taken/not taken ratio)

This also changes the calculation when there is no profile weights: the
existing code inherited the weight from the previous entering block
directly, which does not make much sense if that block has multiple
successors. With this change we assume that the `prevEntering ->
preheader` edge is taken with 1/N probability instead, where N is the
number of successors of `prevEntering`.