Simplify methodtable (#96466)
In particular, simplify access to fields of the MethodTable structure

Historically, this structure had a great many optional fields in it with several mechanisms for identifying them. As we've been working on CoreCLR, we've eliminated many of these excess fields, to the point where we likely have a final list of all expandable fields associated with the MethodTable, and we have space to make access to fields somewhat simpler.

Notably, this is an effort to reduce the number of memory reads and calculations necessary to find the various parts of a MethodTable. With this change, I believe that all data directly associated the MethodTable struct can be accessed with no more than 2 pointer chasing memory accesses and possibly reading 1 set of flags. This should simplify much of the codegen around this structure, although performance is likely unaffected, as most of the slow paths were very well optimized by native PGO technologies.